requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4
alpaca-trade-api==1.3.0
ansiwrap==0.8.4
appdirs==1.4.4
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.22.0
attrs==20.3.0
black==21.4b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dataclasses==0.6
entrypoints==0.3
eventkit==0.8.6
fastapi==0.63.0
greenlet==1.1.1
h11==0.12.0
ib-insync==0.9.65
idna==2.10
importlib-metadata==4.0.1
ipython-genutils==0.2.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.20.1
pandas==1.2.3
papermill==2.3.3
pathspec==0.8.1
praw==7.4.0
prawcore==2.3.0
psycopg2==2.9.1
pydantic==1.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.23
starlette==0.13.6
tenacity==7.0.0
textwrap3==0.9.2
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.4
uvicorn==0.13.4
websocket-client==0.58.0
websockets==8.1
yarl==1.6.3
zipp==3.4.1
